+++ Package.swift
@@ -1,12 +1,12 @@
 // swift-tools-version:5.8
 import PackageDescription
 // The release tag for this version of ApproovURLSession
-let releaseTAG = "3.5.6"
+let releaseTAG = "3.5.7"
 // SDK package version (used for both iOS and watchOS)
 let sdkVersion: Version = "3.5.3"
 
 let package = Package(
-    name: "ApproovURLSession",
+    name: "ApproovURLSessionPackage",
     platforms: [
         .iOS(.v11),
         .watchOS(.v9)
@@ -14,13 +14,13 @@ let package = Package(
     products: [
         // Combined library for iOS and watchOS
         .library(
-            name: "ApproovURLSession",
-            targets: ["ApproovURLSession"]
+            name: "ApproovURLSessionPackage",
+            targets: ["ApproovURLSessionPackage"]
         ),
         .library(
             name: "ApproovURLSessionDynamic",
             type: .dynamic,
-            targets: ["ApproovURLSession"]
+            targets: ["ApproovURLSessionPackage"]
         )
     ],
     dependencies: [
@@ -29,7 +29,7 @@ let package = Package(
     ],
     targets: [
         .target(
-            name: "ApproovURLSession",
+            name: "ApproovURLSessionPackage",
             dependencies: [
                 .product(name: "Approov", package: "approov-ios-sdk"),
                 .product(name: "RawStructuredFieldValues", package: "swift-http-structured-headers")