+++ Sources/ApproovURLSession/ApproovServiceMutator.swift
@@ -195,11 +195,8 @@ public extension ApproovServiceMutator {
         case .noNetwork,
              .poorNetwork,
              .mitmDetected:
-            if !ApproovService.getProceedOnNetworkFailure() {
-                throw ApproovError.networkingError(message: "Approov token fetch for \(url): " +
-                                                   Approov.string(from: status))
-            }
-            return false
+            throw ApproovError.networkingError(message: "Approov token fetch for \(url): " +
+                                               Approov.string(from: status))
         case .noApproovService,
              .unknownURL,
              .unprotectedURL:
@@ -223,11 +220,8 @@ public extension ApproovServiceMutator {
         case .noNetwork,
              .poorNetwork,
              .mitmDetected:
-            if !ApproovService.getProceedOnNetworkFailure() {
-                throw ApproovError.networkingError(message: "Header substitution for \(header): " +
-                                                   Approov.string(from: status))
-            }
-            return false
+            throw ApproovError.networkingError(message: "Header substitution for \(header): " +
+                                               Approov.string(from: status))
         case .unknownKey:
             return false
         default:
@@ -249,11 +243,8 @@ public extension ApproovServiceMutator {
         case .noNetwork,
              .poorNetwork,
              .mitmDetected:
-            if !ApproovService.getProceedOnNetworkFailure() {
-                throw ApproovError.networkingError(message: "Query parameter substitution for \(queryKey): " +
-                                                   Approov.string(from: status))
-            }
-            return false
+            throw ApproovError.networkingError(message: "Query parameter substitution for \(queryKey): " +
+                                               Approov.string(from: status))
         case .unknownKey:
             return false
         default: